+++ buildSrc/src/main/kotlin/Deps.kt
@@ -1,10 +1,11 @@
 object Deps {
+    const val desugaring = "com.android.tools:desugar_jdk_libs:1.1.5"
     const val annotations =  "androidx.annotation:annotation:1.2.0"
     const val rxjava3 =  "io.reactivex.rxjava3:rxjava:3.0.13"
     const val rxandroid =  "io.reactivex.rxjava3:rxandroid:3.0.0"
     const val rxbinding =  "com.jakewharton.rxbinding4:rxbinding:4.0.0"
     const val junit =  "junit:junit:4.12"
-    const val assertj =  "org.assertj:assertj-core:3.19.0"
-    const val robolectric =  "org.robolectric:robolectric:4.5.1"
+    const val assertj =  "org.assertj:assertj-core:3.20.2"
+    const val robolectric =  "org.robolectric:robolectric:4.6.1"
     const val testcore =  "androidx.test:core:1.3.0"
 }