+++ buildSrc/src/main/kotlin/Versions.kt
@@ -5,6 +5,5 @@ object Versions {
     const val targetSdkVersion = 30
     const val compileSdkVersion = 30
     const val buildToolsVersion = "30.0.2"
-    val javaVersion = JavaVersion.VERSION_1_7
-    val java8Version = JavaVersion.VERSION_1_8
+    val javaVersion = JavaVersion.VERSION_1_8
 }